requirements.txt
openai>=1.40.0
pypdf>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.1
pdf2image>=1.17.0
pytesseract>=0.3.10
